# shared/openapi.yaml
openapi: 3.0.3
info:
  title: AlpRoutes API
  version: 0.1.0
  description: |
    Единый API для Android- и веб-клиента каталога альпинистских маршрутов.
    Этот файл — источник истины: клиентские модели генерируются из него.

    ## Общие правила

    * **Геометрия** — GeoJSON, EPSG:4326, порядок координат `[lon, lat]` или `[lon, lat, ele]`.
      `bbox` в запросах — `minLon,minLat,maxLon,maxLat`.
    * **Пагинация** — у любого списка. Курсорная: `cursor` + `limit` (1–100, по умолчанию 20),
      в ответе `nextCursor` (null — страниц больше нет). Курсор непрозрачный.
    * **Мультиязычность** — тексты каталога (`LocalizedText`) приходят целиком, со всеми
      переводами: `{"ru": "...", "en": "..."}`. Язык выбирает клиент
      (порядок: язык интерфейса → `ru` → любой имеющийся). Пользовательский контент
      (комментарии, отчёты) не переводится и несёт поле `language`.
    * **Отсутствующие данные** — `null` означает «нет данных». Клиент показывает «нет данных»,
      а не ноль и не прочерк, похожий на значение. Сервер никогда не подставляет оценку.
    * **Категории сложности** — всегда пара `{system, value}`, у маршрута может быть несколько,
      не больше одной на систему. Конвертации между системами в API нет и не будет.
    * **Офлайн-запись.** Всё, что пользователь создаёт (восхождения, комментарии, фото, треки),
      создаётся с `id`, сгенерированным на клиенте (UUID). Повтор того же запроса с тем же `id`
...
    * **Ошибки** — `application/problem+json` (RFC 9457).
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /areas:
    get:
      tags: [areas]
      operationId: listAreas
      summary: Список районов
      description: Без `parentId` и `q` возвращает корневые районы.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: parentId
          in: query
          schema: { type: string, format: uuid }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/AreaType' }
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Bbox'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AreaPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [areas]
      operationId: createArea
      summary: Создать район (модератор)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AreaCreate' }
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Area' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /areas/{areaId}:
    parameters:
      - $ref: '#/components/parameters/AreaId'
    get:
      tags: [areas]
      operationId: getArea
      summary: Район
      security: [{}, { bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Area' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [areas]
      operationId: updateArea
      summary: Изменить район (модератор)
      description: Передаются только изменяемые поля. `null` очищает поле.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AreaUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Area' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /areas/by-slug/{slug}:
    get:
      tags: [areas]
      operationId: getAreaBySlug
      summary: Район по slug (для SSR-страниц)
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Area' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token из `TokenPair`. Живёт недолго, обновляется через `/auth/refresh`.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Непрозрачный курсор из `nextCursor` предыдущей страницы
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SearchQuery:
      name: q
      in: query
      description: Поиск по названию на всех языках (триграммы), минимум 2 символа
      schema: { type: string, minLength: 2, maxLength: 100 }
    Bbox:
      name: bbox
      in: query
      description: '`minLon,minLat,maxLon,maxLat` в EPSG:4326'
      schema: { type: string, example: '42.9,43.0,43.3,43.2' }
    Slug:
      name: slug
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Slug' }
...
    AreaId:
      name: areaId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Нужна аутентификация или токен недействителен
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Недостаточно прав
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: |
        Конфликт: занятый slug, чужой объект с тем же `id`, устаревшая базовая ревизия
        (тогда в `Problem.currentRevisionId` — актуальная) или недопустимый переход статуса.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 9457. Значения `type`: `validation-error`, `unauthorized`, `forbidden`, `not-found`,
        `conflict`, `slug-taken`, `revision-conflict`, `revision-stale`, `invalid-state`,
        `track-not-contiguous`, `payload-too-large`, `internal-error`.
      required: [type, title, status]
      properties:
        type: { type: string, description: 'Машиночитаемый код, например `revision-conflict`' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        currentRevisionId:
          type: string
          format: uuid
          description: Только для `revision-conflict`
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string, example: 'content.grades[0].value' }
        message: { type: string }
    Slug:
      type: string
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      maxLength: 120
      example: dykh-tau-mummery
    LocalizedText:
      type: object
      description: |
        Переводы текста: ключ — код языка ISO 639-1 (`ru`, `en`, ...), значение — непустая строка.
        Хотя бы один перевод.
      minProperties: 1
      additionalProperties: { type: string, minLength: 1 }
      example: { ru: Дых-Тау, en: Dykh-Tau }
...
    Position:
      type: array
      description: '`[lon, lat]` или `[lon, lat, ele]`, высота в метрах. Используется для треков и контуров.'
      minItems: 2
      maxItems: 3
      items: { type: number, format: double }
      example: [43.1003, 43.0342, 5205]
    Position2D:
      type: array
      description: '`[lon, lat]`. Высоты у точек и нарисованных линий не хранятся.'
      minItems: 2
      maxItems: 2
      items: { type: number, format: double }
      example: [43.1003, 43.0342]
    GeoJsonPoint:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Point] }
        coordinates: { $ref: '#/components/schemas/Position2D' }
...
    GeoJsonMultiPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [MultiPolygon] }
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              minItems: 4
              items: { $ref: '#/components/schemas/Position' }

    # --------------------------------------------------------------- auth
...
    AreaType:
      type: string
      enum: [mountain_system, region, ridge, massif, summit, other]
    ContentStatus:
      type: string
      enum: [draft, published, hidden]
    AreaRef:
      type: object
      required: [id, slug, type, name]
      properties:
        id: { type: string, format: uuid }
        slug: { $ref: '#/components/schemas/Slug' }
        type: { $ref: '#/components/schemas/AreaType' }
        name: { $ref: '#/components/schemas/LocalizedText' }
    AreaSummary:
      allOf:
        - $ref: '#/components/schemas/AreaRef'
        - type: object
          required: [status, updatedAt]
          properties:
            parentId: { type: string, format: uuid, nullable: true }
            center: { $ref: '#/components/schemas/GeoJsonPoint' }
            elevationM:
              type: integer
              nullable: true
              description: Только для вершин. null — нет данных.
            status: { $ref: '#/components/schemas/ContentStatus' }
            childCount: { type: integer }
            routeCount: { type: integer, description: Опубликованные маршруты, включая вложенные районы }
            updatedAt: { type: string, format: date-time }
    Area:
      allOf:
        - $ref: '#/components/schemas/AreaSummary'
        - type: object
          required: [ancestors]
          properties:
            description: { $ref: '#/components/schemas/LocalizedText' }
            boundary: { $ref: '#/components/schemas/GeoJsonMultiPolygon' }
            ancestors:
              type: array
              description: Хлебные крошки от корня к родителю
              items: { $ref: '#/components/schemas/AreaRef' }
    AreaCreate:
      type: object
      required: [type, slug, name]
      properties:
        parentId: { type: string, format: uuid, nullable: true }
        type: { $ref: '#/components/schemas/AreaType' }
        slug: { $ref: '#/components/schemas/Slug' }
        name: { $ref: '#/components/schemas/LocalizedText' }
        description: { $ref: '#/components/schemas/LocalizedText' }
        center: { $ref: '#/components/schemas/GeoJsonPoint' }
        elevationM: { type: integer, nullable: true, minimum: -500, maximum: 9000 }
        boundary: { $ref: '#/components/schemas/GeoJsonMultiPolygon' }
        status: { $ref: '#/components/schemas/ContentStatus' }
    AreaUpdate:
      type: object
      description: Только изменяемые поля
      properties:
        parentId: { type: string, format: uuid, nullable: true }
        type: { $ref: '#/components/schemas/AreaType' }
        slug: { $ref: '#/components/schemas/Slug' }
        name: { $ref: '#/components/schemas/LocalizedText' }
        description: { $ref: '#/components/schemas/LocalizedText' }
        center: { $ref: '#/components/schemas/GeoJsonPoint' }
        elevationM: { type: integer, nullable: true, minimum: -500, maximum: 9000 }
        boundary: { $ref: '#/components/schemas/GeoJsonMultiPolygon' }
        status: { $ref: '#/components/schemas/ContentStatus' }
    AreaPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AreaSummary' }
        nextCursor: { type: string, nullable: true }

    # ------------------------------------------------------------- routes
...